rename jobs.ts -> useJobs.ts
