+++ src/sections/detailsDrawer/RiskDrawer.tsx
@@ -18,10 +18,10 @@ import { Tooltip } from '@/components/Tooltip';
 import { DetailsListContainer } from '@/components/ui/DetailsListContainer';
 import { RiskDropdown } from '@/components/ui/RiskDropdown';
 import { useMy } from '@/hooks';
-import { useReRunJob } from '@/hooks/jobs';
 import { useGetFile, useUploadFile } from '@/hooks/useFiles';
 import { useGenericSearch } from '@/hooks/useGenericSearch';
 import { useInterval } from '@/hooks/useInterval';
+import { useReRunJob } from '@/hooks/useJobs';
 import { useReportRisk, useUpdateRisk } from '@/hooks/useRisks';
 import { useSearchParams } from '@/hooks/useSearchParams';
 import { useSearchContext } from '@/state/search';
